requirements.txt
Flask==3.0.3
pymongo==4.7.2
Flask-Login==0.6.3
python-dotenv==1.0.1
bcrypt==4.1.3
dnspython==2.6.1
